# backend/src/rest/v1/openapi.yaml
openapi: 3.0.3
info:
  title: ChatFlow API
  description: A minimal but clean WhatsApp-style messaging app API
  version: 1.0.0
  contact:
    name: ChatFlow Team
servers:
  - url: http://localhost:3002/v1
    description: Development server
  - url: https://chatflow-backend-3w6u4kmniq-ue.a.run.app/v1
    description: Production server (Cloud Run)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        isOnline:
          type: boolean
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details: {}
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details: {}
...
    UserKeyContext:
      type: object
      properties:
        userEmail:
          type: string
          format: email
        keyIds:
          type: object
          properties:
            message:
              type: string
            search:
              type: string
            suggestion:
              type: string
        derivationMethod:
          type: string
        instructions:
          type: object
    KeyVerificationRequest:
      type: object
      required:
        - testData
      properties:
        testData:
          type: string
          description: Plain text to encrypt and verify
        purpose:
          type: string
          enum: [message, search, suggestion]
          default: message
    KeyVerificationResult:
      type: object
      properties:
        verified:
          type: boolean
        originalData:
          type: string
        encryptedData:
          $ref: '#/components/schemas/EncryptedField'
        decryptedData:
          type: string
        match:
          type: boolean
        message:
          type: string

    # Encryption Schemas
    EncryptionMetadata:
      type: object
      description: "Metadata for field-level encryption using AES-256-GCM"
      required:
        - algorithm
        - keyId
        - iv
        - tag
      properties:
        algorithm:
          type: string
          enum: ["AES-256-GCM"]
          description: "Encryption algorithm used"
          example: "AES-256-GCM"
        keyId:
          type: string
          description: "Client-side key identifier for key management"
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "user_key_001"
        iv:
          type: string
          format: base64
          description: "Base64 encoded initialization vector (96-bit for GCM)"
          example: "MTIzNDU2Nzg5MGFiY2RlZg=="
        tag:
          type: string
          format: base64
          description: "Base64 encoded authentication tag (128-bit for GCM)"
          example: "dGFnMTIzNDU2Nzg5MGFiY2Q="
        timestamp:
          type: integer
          format: int64
          description: "Unix timestamp for replay attack prevention"
          example: 1640995200
        nonce:
          type: string
          description: "Unique nonce for each encryption operation"
          example: "unique_nonce_001"
    EncryptedField:
      type: object
      description: "Encrypted field with metadata"
      required:
        - data
        - encryption
      properties:
        data:
          type: string
          format: base64
          description: "Base64 encoded encrypted data"
          example: "eyJhbGciOiJBMjU2R0NNIiwiZW5jIjoiQTI1NkdDTSJ9..."
        encryption:
          $ref: '#/components/schemas/EncryptionMetadata'
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/me/keys/context:
    get:
      tags:
        - Users - Key Context
      summary: Get user's encryption key context
      description: Retrieve user-specific key context for client-side encryption
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's key context for encryption
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserKeyContext'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/keys/verify:
    post:
      tags:
        - Users - Key Context
      summary: Verify client encryption/decryption capability
      description: Test that the client can properly encrypt and decrypt data with their keys
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyVerificationRequest'
      responses:
        '200':
          description: Encryption verification results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/KeyVerificationResult'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Verification error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...